Begin test class exploration
There's a new test_class example for the Teensy 4. It needs a fork
of usb-device to support test device generation. It seems to
communicate with the host-side test harness. Flash the test_class
example to your Teensy 4, and you should see a blinking LED.

To run the host-side test harness, clone the usb-device fork at the
same level as your imxrt-usbd repo. Then, from inside the usb-device
repo, run 'cargo test' to execute the test.

We're not passing right now :P