Time USB polling, pass all tests
Failures are due to timing inconsistencies, caused by the very fast
polling rate. We noticed issues where the test class wouldn't expect
data to mainfest out of nowhere, which could only happen when we're
checking at very, very small intervals. Now, a poll on the device
is much more likely to correlate with a poll on the class.