Add support for USB GPTs; test GPTs on Teensy 4
Each USB peripheral has two dedicated general purpose timers (GPTs).
They're configured through the USB core registers, have a 1us
resolution, and a 24-bit counter. The GPTs can generate interrupts that
are independent of other USB interrupts. This makes GPTs useful for
scheduling USB transfers that are independent of USB traffic or
host-driven events.

This commit adds an API for the GPTs, and access to the GPTs through the
BusAdapter and usb_device APIs. See the documentation for usage.

Tested on the Teensy 4. See the new example that toggles an LED. The
test demonstrates how the GPTs can be used independent of any USB device
I/O.