Keep all unhandled status bits high
If a reset and endpoint event coincide (because you're configuring ITC,
for instance), the first poll call would clear all the bits when
returning early for the reset event. This could lose the endpoint status
bits. This commit updates poll so that it only clears bits that it has
handled. Keeping unhandled bits high allows an interrupt to re-trigger,
or the next poll to observe the event.