Require that the user allocates endpoint state
An EndpointState struct is allocated by the user in a static. The user
can specify the total number of endpoints, but the default value is the
maximum number of endpoints. The user supplies this state to the driver.

A child commit will refactor the Peripherals trait to return raw
pointers to register blocks. Once we do this, we wouldn't be able to use
the existing state module as-is, since we won't be able to distingush
between USB1 and USB2. After this commit, that won't be a problem
anymore.

Users can now optionally reduce the number of supported endpoints to
save some memory, something that wasn't possible before this commit.

Tested using the test_class. All tests passed. Miri seemed OK with the
simple unit tests that check the endpoint allocator.