Update Peripherals trait to provide pointers
The 11xx MCUs have different addresses for USB peripherals. Turns out
the 1010 MCUs also have different addresses. So, it seems we can't
hard-code addresses without introducing some kind of conditional build.

Instead of introducing those feature flags, this commit refactors the
Peripherals trait. Implementers are not responsible for returning
pointers to register blocks. The driver casts those to the register
blocks. Register blocks are consistent across all considered 10xx and
11xx MCUs.

Another approach is to revert the separation from imxrt-ral. We would
depend on imxrt-ral, and users would be responsible for selecting a RAL
feature in their build. This loses flexibility, but we'll be able to use
the imxrt-ral register block and instances.