Show version incompatibility with cortex-m
The previous commits restrict which cortex-m versions a user can build
with. This commit demonstrates the issue. From the root of the repo, run
the build with

    cargo build --manifest-path i-need-min-cm/Cargo.toml

i-need-min-cm represents a closed source user library. For reasons out
of our control, that user really wants to use 0.7.0 of cortex-m.

We _could_ allow imxrt-usbd to build with version 0.7.0. We've shown
that it works. But we're not letting that build happen.