Require imxrt-ral pacakge
imxrt-hal was previously responsible for providing a imxrt-ral
compatibility layer. Now, imxrt-usbd provides the compatibility layer
itself by taking a imxrt-ral dependency. This lets us remove imxrt-usbd
as a imxrt-hal dependency. However, it means this USB driver pulls
in the entire RAL. I'm finding this easier to maintain, even if it's
pulling in more than necessary.